+++ README.md
@@ -3,3 +3,7 @@
 Big things coming
 
 This project was initialized with [create-react-app](https://create-react-app.dev/)
+
+### TODO:
+
+- Figure out how to do tests on `useEffect` in custom hooks
