+++ package.json
@@ -48,9 +48,11 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/react-hooks": "^5.0.3",
     "@types/react-toggle": "^4.0.2",
     "eslint-config-prettier": "^7.2.0",
     "husky": "^5.0.9",
+    "jest-localstorage-mock": "^2.4.6",
     "postcss": "^8.2.6",
     "postcss-cli": "^8.3.1",
     "prettier": "^2.2.1",
