+++ src/modules/homepage/components/theme-switcher-icons/index.ts
@@ -1,0 +1,1 @@
+export * from './theme-switcher-icons';
