+++ src/modules/homepage/components/theme-switcher-icons/theme-switcher-icons.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+
+export const LightIcon = () => <i className="fas fa-sun text-white"></i>;
+export const DarkIcon = () => <i className="fas fa-moon"></i>;
