+++ src/modules/homepage/containers/homepage/index.ts
@@ -1,0 +1,1 @@
+export * from './homepage';
