+++ src/modules/homepage/containers/index.ts
@@ -1,0 +1,1 @@
+export * from './homepage';
