+++ src/modules/shared/hooks/dark-mode/index.ts
@@ -1,0 +1,1 @@
+export * from './use-dark-mode';
