+++ src/modules/shared/hooks/favicon-switcher/index.ts
@@ -1,0 +1,1 @@
+export * from './user-favicon-switcher';
