+++ src/modules/shared/hooks/favicon-switcher/use-favicon-switcher.test.ts
@@ -1,0 +1,7 @@
+describe('useFaviconSwitcher', () => {
+  test('calls the favicon switcher', () => {
+    // TODO: figure out how to trigger useEffect
+
+    expect(true).toBe(true);
+  });
+});
