+++ src/modules/shared/hooks/favicon-switcher/user-favicon-switcher.ts
@@ -1,0 +1,8 @@
+import { updateFavicon } from './../../utilities';
+import { useEffect } from 'react';
+
+export const useFaviconSwitcher: (mode: string) => void = (mode: string) => {
+  useEffect(() => {
+    updateFavicon(mode);
+  });
+};
