+++ src/modules/shared/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './dark-mode';
+export * from './favicon-switcher';
