+++ src/modules/shared/utilities/update-favicon/index.ts
@@ -1,0 +1,1 @@
+export * from './update-favicon';
